# src/main/java/fr/carrefour/biskot/dto/Cart.java
package fr.carrefour.biskot.dto;

import com.fasterxml.jackson.annotation.JsonProperty;
import lombok.Data;

import java.util.ArrayList;
import java.util.List;

@Data
public class Cart {

    private Long id ;
    private List<AddProduct> products = new ArrayList<>();
    @JsonProperty(access = JsonProperty.Access.WRITE_ONLY)
    private Money totalPrice ;

}